requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.13.2
numpy==2.2.6
google-generativeai==0.5.4
groq==1.2.0
python-dotenv==1.0.1
colorama==0.4.6
tqdm==4.66.4
scipy==1.11.4
